Move reusable hCaptcha wrapper code to library.
