+++ packages/hcaptcha/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-09-03
+
+- Adds a reusable native form wrapper to the client entrypoint
+
 ## [1.0.3] - 2026-09-02
 
 - Adds more error cases to the retry loop and updates the readme
