+++ packages/hcaptcha/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gcforms/hcaptcha",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "author": "Canadian Digital Service",
   "license": "MIT",
   "sideEffects": false,
